Allow the user to specify an extra resource path on command line
XcodeGen currently searches for the SettingPresets directory in a
predefined hardcoded set of directories. The user may decide to store
SettingPresets in a non standard location, and in these case it would be
nice to let the user pass this location using a command line argument.

We're adding a --extra-resource-path variadic argument in this patchset to
do that.